+++ CHANGELOG.md
@@ -29,7 +29,7 @@
 
 ## [0.16.2] - 2025-07-16
 * Dask is now an optional dependency, install it with `pip install distributed`
-* Fix bug with gaussianP varstats under 0.5
+* Fix bug with pct2sig and two-sided gaussianP varstats under 0.5
 
 ## [0.16.1] - 2025-07-16
 ### Changed